+++ Pod/Ground.cpp
@@ -1,13 +1,13 @@
 #include "Ground.h"
 
-Ground::Ground(float _x, float _y, float _width, float _height, float _vx, float _vy, float _limitX, float _limitY) : Component(_x, _y, _vx, _vy, _limitX, _limitY)
+Ground::Ground(float _x, float _y, int _width, int _height, float _vx, float _vy, float _limitX, float _limitY) : Component(_x, _y, _vx, _vy, _limitX, _limitY)
 {
 	Component::Component(_x, _y, _vx, _vy, _limitX, _limitY);
 	this->width = _width;
 	this->height = _height;
 }
 
-Ground::Ground(D3DXVECTOR3 _position, float _width, float _height, float _vx, float _vy, float _limitX, float _limitY) : Component(_position, _vx, _vy, _limitX, _limitY)
+Ground::Ground(D3DXVECTOR3 _position, int _width, int _height, float _vx, float _vy, float _limitX, float _limitY) : Component(_position, _vx, _vy, _limitX, _limitY)
 {
 	Component::Component(_position, _vx, _vy, _limitX, _limitY);
 	this->width = _width;
